Add gender field to TenantAdminUser creation
A gender field has been added to the TenantAdminUser creation process. This update includes modification of the User.CreateTenantAdminUser class to now accept a gender input. The data seeding logic for the creation of tenants in the TenantRoleAndUserBuilder.cs file and TenantAppService.cs file has been updated accordingly to accommodate this new field.